SCC0250-CG/proj/t2/requirements.txt
glfw==2.1.0
numpy==1.21.0
Pillow==8.2.0
PyGLM==2.2.0
PyOpenGL==3.1.5
